+++ rusoto/signature/src/signature.rs
@@ -315,7 +315,7 @@ impl SignedRequest {
         self.params
             .insert("X-Amz-Date".into(), current_time_fmted.into());
 
-        self.canonical_query_string = build_canonical_query_string(&self.params, true);
+        self.canonical_query_string = build_canonical_query_string(&self.params);
 
         debug!("canonical_uri: {:?}", self.canonical_uri);
         debug!("canonical_headers: {:?}", canonical_headers);
@@ -380,14 +380,8 @@ impl SignedRequest {
             self.scheme(),
             hostname,
             self.canonical_uri,
-            build_canonical_query_string(&self.params, true)
+            build_canonical_query_string(&self.params)
         )
-    }
-
-    /// Signs the request using Amazon Signature version 4 to verify identity.
-    /// Authorization header uses AWS4-HMAC-SHA256 for signing.
-    pub fn sign(&mut self, creds: &AwsCredentials) {
-        self.sign_with_plus(creds, false)
     }
 
     /// Complement SignedRequest by ensuring the following HTTP headers are set accordingly:
@@ -395,18 +389,9 @@ impl SignedRequest {
     /// - content-type
     /// - content-length (if applicable)
     pub fn complement(&mut self) {
-        self.complement_with_plus(false)
-    }
-
-    /// Complement SignedRequest by ensuring the following HTTP headers are set accordingly:
-    /// - host
-    /// - content-type
-    /// - content-length (if applicable)
-    pub fn complement_with_plus(&mut self, should_treat_plus_literally: bool) {
         // build the canonical request
         self.canonical_uri = self.canonical_path();
-        self.canonical_query_string =
-            build_canonical_query_string(&self.params, should_treat_plus_literally);
+        self.canonical_query_string = build_canonical_query_string(&self.params);
         // Gotta remove and re-add headers since by default they append the value.  If we're following
         // a 307 redirect we end up with Three Stooges in the headers with duplicate values.
         self.remove_header("host");
@@ -430,8 +415,8 @@ impl SignedRequest {
 
     /// Signs the request using Amazon Signature version 4 to verify identity.
     /// Authorization header uses AWS4-HMAC-SHA256 for signing.
-    pub fn sign_with_plus(&mut self, creds: &AwsCredentials, should_treat_plus_literally: bool) {
-        self.complement_with_plus(should_treat_plus_literally);
+    pub fn sign(&mut self, creds: &AwsCredentials) {
+        self.complement();
         let date = PrimitiveDateTime::now();
         self.remove_header("x-amz-date");
         self.add_header("x-amz-date", &date.format("%Y%m%dT%H%M%SZ"));
@@ -693,13 +678,10 @@ fn canonical_uri(path: &str, region: &Region) -> String {
     }
 }
 
-/// Canonicalizes query while iterating through the given parameters.
+/// Canonicalizes query while iterating through the given paramaters
 ///
 /// Read more about it: [HERE](http://docs.aws.amazon.com/AmazonS3/latest/API/sigv4-query-string-auth.html#query-string-auth-v4-signing)
-fn build_canonical_query_string(
-    params: &Params,
-    should_treat_plus_literally: bool,
-) -> String {
+fn build_canonical_query_string(params: &Params) -> String {
     if params.is_empty() {
         return String::new();
     }
@@ -709,19 +691,11 @@ fn build_canonical_query_string(
         if !output.is_empty() {
             output.push_str("&");
         }
-        if should_treat_plus_literally {
-            output.push_str(&encode_uri_strict(&key));
-        } else {
-            output.push_str(&encode_uri_strict(&key.replace("+", " ")));
-        }
+        output.push_str(&encode_uri_strict(&key));
         output.push_str("=");
 
         if let Some(ref unwrapped_val) = *val {
-            if should_treat_plus_literally {
-                output.push_str(&encode_uri_strict(&unwrapped_val));
-            } else {
-                output.push_str(&encode_uri_strict(&unwrapped_val.replace("+", " ")));
-            }
+            output.push_str(&encode_uri_strict(&unwrapped_val));
         }
     }
 
@@ -906,12 +880,8 @@ mod tests {
         let mut params = Params::new();
         for code in start..end {
             params.insert("k".to_owned(), Some((code as char).to_string()));
-            let enc = build_canonical_query_string(&params, false);
-            let expected = if (code as char) == '+' {
-                "k=%20".to_owned()
-            } else {
-                format!("k=%{:02X}", code)
-            };
+            let enc = build_canonical_query_string(&params);
+            let expected = format!("k=%{:02X}", code);
             assert_eq!(expected, enc);
         }
     }
@@ -926,8 +896,8 @@ mod tests {
         request.add_param("arg1%7B", "arg1%7B");
         request.add_param("arg2%7B+%2B", "+%2B");
         assert_eq!(
-            super::build_canonical_query_string(&request.params, false),
-            "arg1%257B=arg1%257B&arg2%257B%20%252B=%20%252B"
+            super::build_canonical_query_string(&request.params),
+            "arg1%257B=arg1%257B&arg2%257B%2B%252B=%2B%252B"
         );
         assert_eq!(
             super::canonical_uri(&request.path, &Region::default()),
@@ -946,7 +916,7 @@ mod tests {
             "key:with@funny&characters",
             "value with/funny%characters/Рускии",
         );
-        let canonical_query_string = super::build_canonical_query_string(&request.params, false);
+        let canonical_query_string = super::build_canonical_query_string(&request.params);
         assert_eq!("key%3Awith%40funny%26characters=value%20with%2Ffunny%25characters%2F%D0%A0%D1%83%D1%81%D0%BA%D0%B8%D0%B8",
                    canonical_query_string);
         let canonical_uri_string = super::canonical_uri(&request.path, &Region::default());
@@ -959,7 +929,7 @@ mod tests {
     fn query_string_literal_plus() {
         let mut params = Params::new();
         params.insert("key".into(), Some("val+ue".into()));
-        let encoded = build_canonical_query_string(&params, true);
+        let encoded = build_canonical_query_string(&params);
         assert_eq!("key=val%2Bue", encoded);
     }